% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_censustable_dict.R
\docType{data}
\name{censustable_dict}
\alias{censustable_dict}
\title{All tables in data files in census 2010 summary file 1 with urban/rural update}
\format{data.table}
\usage{
data("census_table")
}
\description{
This data serves as a dictionary for data files, table numbers, table
description, and table universe.
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
all across chapter 6.
}
\keyword{datasets}
