% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{dict_acs1_geocomponent}
\alias{dict_acs1_geocomponent}
\title{List of geographic components used in ACS 1 year surverys}
\format{A data.table with 28 rows and 9 variables:
\describe{
  \item{code}{code for the geocomponent, such as "01" and "M3"}
  \item{geo_component}{description of the geographic component}
  \item{state_2009_to_now}{wheather a geocomponent available in state files since 2009}
  \item{state_2007_2008}{wheather a geocomponent available in state files in 2007 and 2008}
  \item{state_2005_2006}{wheather a geocomponent available in state files in 2005 and 2006}
  \item{US_2009_to_now}{wheather a geocomponent available in national files since 2009}
  \item{US_2007_2008}{wheather a geocomponent available in national files in 2007 and 2008}
  \item{US_2006}{wheather a geocomponent available in national files in 2006}
  \item{US_2005}{wheather a geocomponent available in national files in 2005}

}}
\usage{
dict_acs1_geocomponent
}
\description{
List of geographic components used in ACS 1 year surverys
}
\keyword{datasets}
