% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{dict_acs1_summarylevel}
\alias{dict_acs1_summarylevel}
\title{List of summary levels used in ACS 5 year surverys}
\format{A data.table with 87 rows and 6 variables
\describe{
  \item{code}{code of summary level}
  \item{summary_level}{description of summary level}
  \item{state_acs1}{wheather a summary level available in 1-year survey state files}
  \item{US_acs1}{wheather a summary level available in 1-year survey national files}
  \item{state_acs5}{wheather a summary level available in 5-year survey state files}
  \item{US_acs5}{wheather a summary level available in 5-year survey national files}
}}
\usage{
dict_acs1_summarylevel
}
\description{
List of summary levels used in ACS 5 year surverys
}
\keyword{datasets}
