% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{dict_acs1_table}
\alias{dict_acs1_table}
\title{List of summary levels used in ACS 5 year surverys}
\format{A data.table with 1324 rows and 6 variables:
\describe{
  \item{table_number}{table number such as "C27013"}
  \item{table_name}{description of the table}
  \item{acs5_2015}{table available in 2015 ACS 5-year survey}
  \item{acs1_2016}{table available in 2016 ACS 1-year survey}
  \item{acs1_2015}{table available in 2015 ACS 1-year survey}
  \item{acs1_2014}{table available in 2014 ACS 1-year survey}
}}
\usage{
dict_acs1_table
}
\description{
List of summary levels used in ACS 5 year surverys
}
\keyword{datasets}
