% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{dict_acs5_table}
\alias{dict_acs5_table}
\title{List of summary levels used in ACS 5 year surverys}
\format{
A data.table with 1174 rows and 14 variables:
\describe{
  \item{table_number}{table number such as "C27013"}
  \item{table_name}{description of the table}
  \item{acs5_2019}{whether the table is available in 2019}
  \item{acs5_2018}{whether the table is available in 2018}
  \item{acs5_2017}{whether the table is available in 2017}
  \item{acs5_2016}{whether the table is available in 2016}
  \item{acs5_2015}{whether the table is available in 2015}
  \item{acs5_2014}{whether the table is available in 2014}
  \item{acs5_2013}{whether the table is available in 2013}
  \item{acs5_2012}{whether the table is available in 2012}
  \item{acs5_2011}{whether the table is available in 2011}
  \item{acs5_2010}{whether the table is available in 2010}
  \item{acs5_2009}{whether the table is available in 2009}
  \item{universe}{universe of the table}
}
}
\usage{
dict_acs5_table
}
\description{
List of summary levels used in ACS 5 year surverys
}
\keyword{datasets}
