% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{dict_acs_geocomponent}
\alias{dict_acs_geocomponent}
\title{List of geographic components used in American Community Survey}
\format{A data.table with 19 rows and 6 variables:
\describe{
  \item{code}{code for the geocomponent, such as "01" and "M3"}
  \item{geo_component}{description of the geographic component}
  \item{state_acs1}{wheather a geocomponent available in 1-year survey state files}
  \item{US_acs1}{wheather a geocomponent available in 1-year survey national files}
  \item{state_acs5}{wheather a geocomponent available in 5-year survey state files}
  \item{US_acs5}{wheather a geocomponent available in 5-year survey national files}
}}
\source{
generated from lookup files
}
\usage{
dict_acs_geocomponent
}
\description{
This dataset contains the geographic components and codes used in
1-year and 5-year surveys. Search geographic components
with function \code{\link{search_geoheaders}}.
}
\keyword{datasets}
