% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{dict_acs_geoheader_2010}
\alias{dict_acs_geoheader_2010}
\title{List of geographic headers used in 2010 ACS 1 and 5 year surveys}
\format{A data.table with 53 rows and 4 variables
\describe{
  \item{reference}{reference of the geoheader}
  \item{field}{description of the geoheader}
  \item{start}{starting position of the geoheader in geography file}
  \item{end}{ending position of the geoheader in geography file}
}}
\source{
2016 ACS Summary File
\href{https://www2.census.gov/programs-surveys/acs/summary_file/2010/documentation/5_year/ACS_2006-2010_SF_Tech_Doc.pdf?#}{technical documentation},
page 11.
}
\usage{
dict_acs_geoheader_2010
}
\description{
List of geographic headers used in 2010 ACS 1 and 5 year surveys
}
\keyword{datasets}
