% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{dict_acs_summarylevel}
\alias{dict_acs_summarylevel}
\title{Summary levels available in American Community Survey}
\format{A data.table with 87 rows and 6 variables
\describe{
  \item{code}{code of summary level}
  \item{summary_level}{description of summary level}
  \item{state_acs1}{wheather a summary level available in 1-year survey state files}
  \item{US_acs1}{wheather a summary level available in 1-year survey national files}
  \item{state_acs5}{wheather a summary level available in 5-year survey state files}
  \item{US_acs5}{wheather a summary level available in 5-year survey national files}
}}
\source{
generated from lookup datasets
}
\usage{
dict_acs_summarylevel
}
\description{
This data contains summary levels and codes used in 1-year and
5-year survey. Search with function \code{\link{search_summarylevels}}.
}
\keyword{datasets}
