% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{dict_acs_table}
\alias{dict_acs_table}
\title{Complete list of ACS tables}
\format{A data.table with 1324 rows and 6 variables:
\describe{
  \item{table_number}{table number such as "C27013"}
  \item{table_name}{description of the table}
  \item{acs5_2015}{table available in 2015 ACS 5-year survey}
  \item{acs1_2016}{table available in 2016 ACS 1-year survey}
  \item{acs1_2015}{table available in 2015 ACS 1-year survey}
  \item{acs1_2014}{table available in 2014 ACS 1-year survey}
}}
\source{
generated from lookup datasets.
}
\usage{
dict_acs_table
}
\description{
This dataset contains all tables in 1-year and 5-year survey.
Note that 5-year survey in 2014 and 2015 has the identical tables.
}
\keyword{datasets}
