% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R, R/data_decennial.R
\docType{data}
\name{dict_all_geocomponent_2000}
\alias{dict_all_geocomponent_2000}
\title{List of all geographic components, 2000 version}
\format{A data.table with 99 rows and 2 variables:
\describe{
  \item{code}{code for the geocomponent, such as "01" and "M3"}
  \item{geo_component}{description of the geographic component}
}}
\source{
2000 Census Summary File 1
\href{https://www.census.gov/prod/cen2000/doc/sf1.pdf}{technical documentation}
page 7-15
}
\usage{
dict_all_geocomponent_2000

dict_all_geocomponent_2000
}
\description{
This dataset contains all available geographic components and codes.
}
\keyword{datasets}
