% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_census.R
\docType{data}
\name{dict_census_geoheader}
\alias{dict_census_geoheader}
\title{List of geographic headers}
\format{A data.table with 101 rows and 4 variables
\describe{
  \item{reference}{reference of the geoheader record}
  \item{field}{description of the geoheader record field}
  \item{start}{starting position of the geoheader in the record}
  \item{end}{ending position of the geoheader in the record}
}}
\source{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 2-8
}
\usage{
dict_census_geoheader
}
\description{
This dataset has the complete list of geographic header
references and their discription used in Census 2010 summary file 1 (with
urban/rural update). Search the dataset with function \code{\link{search_geoheaders}}.
}
\keyword{datasets}
