% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_census.R
\docType{data}
\name{dict_census_summarylevel}
\alias{dict_census_summarylevel}
\title{Summary levels available in Census 2010}
\format{A data.table with 165 rows and 4 variables
\describe{
  \item{code}{code of summary level}
  \item{summary_level}{description of summary level}
  \item{in_state_file}{wheather the summary level available in state files}
  \item{in_US_file}{wheather the summary level available in national files}
}}
\source{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 4-16 state summary file with urban/rural update
}
\usage{
dict_census_summarylevel
}
\description{
This data contains summary levels and codes used in census 2010
 summary file 1 (with urban/rural update). Search with function \code{\link{search_summarylevels}}.
}
\keyword{datasets}
