% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_countyfips_dict.R
\docType{data}
\name{dict_countyfips}
\alias{dict_countyfips}
\title{The fips of all counties in the US}
\format{data.table}
\usage{
data("dict_countyfips")
}
\description{
This data serves as a dictionary for county fips
}
\references{
\href{https://www2.census.gov/geo/docs/reference/codes/files/national_county.txt}{list of all country fips}
}
\keyword{datasets}
