% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_decennial.R
\docType{data}
\name{dict_decennial_table_2010}
\alias{dict_decennial_table_2010}
\title{Complete list of 2010 census tables}
\format{
A data.table with 333 rows and 4 variables:
\describe{
  \item{table_number}{table number such as "H1", "PCT22G"}
  \item{table_name}{description of the table}
  \item{universe}{universe of the data}
  \item{table_ref}{reference code such as "H0010", "PCT022G"}
}
}
\source{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
chapter 5.
}
\usage{
dict_decennial_table_2010
}
\description{
This dataset contains all census tables in census 2010 summary
file 1 (with urban/rural update).
}
\keyword{datasets}
