% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fips_dict.R
\docType{data}
\name{dict_fips}
\alias{dict_fips}
\title{The fips of all counties in the US}
\format{data.table}
\usage{
data("dict_fips")
}
\description{
This data serves as a dictionary for county fips
}
\references{
\href{https://www.census.gov/geographies/reference-files/2016/demo/popest/2016-fips.html}{list of all fips as of 2016}
}
\keyword{datasets}
