% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geocomponent_dict.R
\docType{data}
\name{dict_geocomp}
\alias{dict_geocomp}
\title{Geographic components and codes}
\format{data.table}
\usage{
data("dict_geocomp")
}
\description{
This data serves as a dictionary for geographic component codes and their
discription.
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 6-15
}
\keyword{datasets}
