% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geoheader_dict.R
\docType{data}
\name{dict_geoheader}
\alias{dict_geoheader}
\title{Geographic Header Record — Summary File 1}
\format{data.table}
\usage{
data("dict_geoheader")
}
\description{
This data serves as a dictionary for geographic header references and their
discription.
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 2-8
}
\keyword{datasets}
