% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summarylevel_US_dict.R
\docType{data}
\name{dict_summarylevel_US}
\alias{dict_summarylevel_US}
\title{Summary levels and codes for national files}
\format{A data.table with 77 rows and 2 variables:
\describe{
  \item{code}{code of summary level}
  \item{summary_level}{description of summary level}
}}
\usage{
data("dict_summarylevel_US")
}
\description{
This data contains summary levels and codes used in national files of
 summary file 1 (with urban/rural update). Search with function \code{\link{search_sumlev_US}}.
 This dataset is different from that of state files, \code{\link{dict_summarylevel}},
 which is searched with \code{\link{search_sumlev}}.
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 4-25 state summary file with urban/rural update
}
\keyword{datasets}
