% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_census.R
\name{download_census}
\alias{download_census}
\title{download census data}
\usage{
download_census(survey, year, states = c(states_DC, "US", "PR"))
}
\arguments{
\item{survey}{Which survey to download from, "decennial", "acs5year", ,
"acs1year", or "redistricting".}

\item{year}{year or ending year of the survey}

\item{states}{vector of abbreviations of states such as c("MA", "RI")}
}
\description{
Download decennial census and ACS 5-year and 1-year data from
United States Census bureau. It also download
generated data from Census 2010 if not exist.
}
