% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_census.R
\name{download_generated_data}
\alias{download_generated_data}
\title{Download data generated from Census 2010}
\usage{
download_generated_data()
}
\description{
This function downloads data generated from Census 2010
from Census 2010.
}
