% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_packages.R
\name{dummy_function}
\alias{dummy_function}
\title{Import packages and functions}
\usage{
dummy_function()
}
\description{
The purpose of this file is to import all packages and
functions used in #' the package. Have to create a dummy function to make
the importing work.
}
