% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_from_census2010.R
\name{generate_census_data}
\alias{generate_census_data}
\title{Generate additional geographic data from Census 2010.}
\usage{
generate_census_data(states = c(states_DC, "PR", "US"))
}
\arguments{
\item{states}{vector of abbreviations of states, such as c("MA", "RI")}
}
\description{
Decennial census has the most complete geographic header records,
which can fill in many missing values in ACS data. This function is to generate
such datasets. It creates a sub directory "generared_data/" under path_to_census
if not exsist and store generated data.
}
