% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_geoid_coordinate.R
\name{generate_geoidcoord}
\alias{generate_geoidcoord}
\title{Generate coordinate for selected GEOIDs in ACS 5-year survey}
\usage{
generate_geoidcoord()
}
\description{
It creates a sub directory "generared_data/" under path_to_census
if not exsist and a file called geoid_coord.csv. The coordinates are generated
from census 2010 summary file 1 (with urban/rural update)
}
