% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geo_dict.R
\docType{data}
\name{geo_dict}
\alias{geo_dict}
\title{Geographic Header Record — Summary File 1}
\format{data.table}
\usage{
data("geo_header_dictionary")
}
\description{
This data serves as a dictionary for geographic header references and their
discription.
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 2-8
}
\keyword{datasets}
