% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_all_acs_lookup.R
\docType{data}
\name{lookup_acs1year_2016}
\alias{lookup_acs1year_2016}
\title{ACS 1-year 2016 file segment and table lookup data}
\format{A data.table with 4 variables
\describe{
  \item{file_segment}{sequence number of segment data files, from 1 to 166}
  \item{table_content}{description of content in a census table}
  \item{reference}{reference of the field, such as "B01001_002"}
  \item{table_number}{table number such as "B01001"}
  \item{table_name}{description of table, which has many table_content}
  \item{universe}{the unit of the census data}
}}
\source{
Check for each year of ACS 1-year and 5-year
\href{https://www.census.gov/programs-surveys/acs/technical-documentation/summary-file-documentation.2015.html}{Sequence Number/Table Number Lookup File}.
}
\usage{
lookup_acs1year_2016
}
\description{
ACS 1-year 2016 file segment and table lookup data
}
\keyword{datasets}
