% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geoheader_and_data.R
\name{read_2010census}
\alias{read_2010census}
\title{Read selected geoheader and table content}
\usage{
read_2010census(path_to_census, state, geoheaders, table_contents,
  show_progress = TRUE)
}
\arguments{
\item{path_to_census}{path to the directory holding downloaded
census 2010 summary file 1 with urban/rural update}

\item{state}{abbreviation of a state, for example "IN" for "Indiana"}

\item{geoheaders}{vector of references of selected geographci headers to be included in the return}

\item{table_contents}{selected references of contents in census tables}

\item{show_progress}{show progress of reading if TRUE. Turn off if FALSE, which
is useful in RMarkdown output.}
}
\value{
A data.table of selected geoheaders and table contents, with LOGRECNO
as key.
}
\description{
Read selected geoheader and table content
}
\examples{
\dontrun{
aaa <- read_2010census("~/dropbox_datasets/US_2010_census/", "RI",
                       c("STATE", "NAME", "SUMLEV"),
                       c("PCT012F139", "PCT012F138", "P0150008", "H0070016"))
}

}
