% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datafile.R
\name{read_2010fileseg}
\alias{read_2010fileseg}
\title{Read data file}
\usage{
read_2010fileseg(path_to_census, state, file_seg)
}
\arguments{
\item{path_to_census}{path to the directory of downloaded census 2010 data}

\item{state}{abbrivation of a state, for example, "IN" for Indiana}

\item{file_seg}{the number of the file, for example, 5 for file 05}
}
\value{
A data.table of all data in the data file
}
\description{
Read an entire data file in a state and assign correct reference to each column
}
\examples{
\dontrun{
path <- your_local_path_to_census_data
aaa <- read_datafile(path, "RI", 2)
}

}
\seealso{
\code{\link{dict_datafile}}
}
