% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datafile.R
\name{read_2010fileseg_}
\alias{read_2010fileseg_}
\title{Read a data file of a state (internal)}
\usage{
read_2010fileseg_(state, file_seg)
}
\arguments{
\item{state}{abbrivation of a state, for example, "IN" for Indiana.}

\item{file_seg}{the number of the data file, for example, 5 for file 05.}
}
\value{
A data.table of all data in the data file. Its columns are the references
of LOGRECNO and tables contents in the file. LOGRECNO is its key.
}
\description{
Read an entire data file in a state and assign correct reference
to each column. Search with \code{\link{search_datafile}} to find table contents
in a data file.
}
\examples{
\dontrun{
# read Rhode Island data file 02
ri_2 <- read_datafile("your_local_path_to_census_data", "RI", 2)
}

}
\seealso{
\code{\link{dict_datafile}}
}
