% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geographic_header_record_file.R
\name{read_2010geoheader}
\alias{read_2010geoheader}
\title{Read geographic header record file of a state}
\usage{
read_2010geoheader(path_to_census, state, geoheaders = NULL,
  show_progress = TRUE)
}
\arguments{
\item{path_to_census}{path to the directory holding downloaded
census 2010 summary file 1 with urban/rural update.}

\item{state}{abbreviation of a state, for example "IN" for "Indiana".}

\item{show_progress}{show progress of reading if TRUE. Turn off if FALSE, which
is useful in RMarkdown output.}

\item{references}{vector of references of selected geographic headers to be included in the return}
}
\value{
data.table whose columns are logical record number, summary level,
geographic compoenent, and selected references. LOGRECNO serves as
the key.
}
\description{
Read geographic header record file of a state and return logical record number
and selected geographic headers. To find geographic headers, browse
\code{\link{dict_geoheader}} or search with \code{\link{search_geoheader}}.
}
\examples{
\dontrun{
# read selected geographic headers of Rhode Island
ri  <- read_2010geoheader("your_local_path_to_census_data", "RI",
                      c("NAME", "SUMLEV", "INTPTLAT", "INTPTLON"))
}


}
