% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_censustable.R
\name{read_2010table}
\alias{read_2010table}
\title{Read a census table of a state}
\usage{
read_2010table(state, table_number)
}
\arguments{
\item{state}{abbrivation of a state, such as "IN" for Indiana}

\item{table_number}{table number such as "P3" and "PCT11A"}

\item{path_to_census}{path to the directory holding downloaded census data,
under which are sub-folders of each state.}
}
\value{
A data.table with of the selected census table. LOGRECNO is the key.
}
\description{
Read a census table of a state. To find table of interest, search with function
\code{\link{search_table}} or browse the dictionary \code{\link{dict_censustable}}.
}
\examples{
\dontrun{
# read table P11 of Rhode Island
ri <- read_2010table("your_local_path_to_census_folder", "RI", "P11")
}


}
