% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acs5year.R
\name{read_acs5location}
\alias{read_acs5location}
\title{Read ACS 5-year survey by location}
\usage{
read_acs5location(year, states, table_contents, locations,
  summary_level = "block group", geo_comp = "*", with_margin = FALSE,
  show_progress = TRUE)
}
\arguments{
\item{year}{end year of the 5-year survey}

\item{states}{vector of abbreviations of states such as c("MA", "RI")}

\item{table_contents}{vector of reference of available table contents}

\item{locations}{For metro area, in the format like "New York metro".
 For county, city, or town, must use the exact name as those in
 \code{\link{dict_fips}} in the format like "kent county, RI",
"Boston city, MA", and "Lincoln town, RI". And special examples like
"Salt Lake City city, UT" must keep the "city" after "City".}

\item{summary_level}{summary level like "050"}

\item{geo_comp}{geographic component such as "00", "01", and "43"}

\item{with_margin}{read also margin of error in addition to estimate}

\item{show_progress}{whether to show progress in fread()n}
}
\value{
A data.table of selected data.
}
\description{
Read ACS 5-year survey of selected geography, estimate, and margin
of selected states.
}
\examples{
\dontrun{
# read a single location and one table content
aaa <- read_acs5location(2015, "RI",
                         table_contents = "B01001_009",
                         locations = "providence city, ri",
                         summary_level = "block group")


# read multiple locations
bbb <- read_acs5location(2015, c("ut", "ri"),
                      table_contents = c("B01001_009", "B00001_001", "B10001_002"),
                      locations = c("Lincoln Town, RI",
                                      "Kent county, RI",
                                      "Salt Lake City city, UT",
                                      "Salt Lake metro",
                                      "Providence city, RI"),
                      summary_level = "block group")
}


}
