% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acs5year.R
\name{read_acs5year}
\alias{read_acs5year}
\title{Read ACS 5-year survey}
\usage{
read_acs5year(year, states, table_contents = NULL, locations = NULL,
  geo_headers = NULL, summary_level = "*", geo_comp = "*",
  with_margin = FALSE, with_coord = TRUE, with_population = TRUE,
  with_acsgeoheaders = FALSE, show_progress = TRUE)
}
\arguments{
\item{year}{end year of the 5-year survey}

\item{states}{vector of abbreviations of states such as c("MA", "RI")}

\item{table_contents}{vector of reference of available table contents}

\item{locations}{for county, city, or town, in the format like "kent county, RI",
"Boston city, MA", "Lincoln town, RI". For metro area, in the format
like "New York metro".}

\item{geo_headers}{vector of geographic headers like c("NAME", "PLACE")}

\item{summary_level}{summary level like "050"}

\item{geo_comp}{geographic component such as "00", "01", and "43"}

\item{with_margin}{read also margin of error in addition to estimate}

\item{with_coord}{whether to include longitude and latitude in return}

\item{with_population}{whether to include total population in return}

\item{with_acsgeoheaders}{whether to include geographic headers from ACS data}

\item{show_progress}{whether to show progress in fread()n}
}
\value{
A data.table of selected data.
}
\description{
Read ACS 5-year survey of selected geography, estimate, and margin
of selected states.
}
\examples{
\dontrun{
aaa <- read_acs5year(2015, c("ut", "ri"),
                      table_contents = c("B01001_009", "B00001_001", "B10001_002"),
                      geo_headers = c("place = ut62360",  # place in UT
                                      "county = ut001",
                                      "place = ri14140",    # place in UT or RI
                                      "cousub = ri41500", # county sub in RI
                                      "cbsa = 39300"),
                      summary_level = "block group")
}


}
