% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acs_estimate_and_margin.R
\name{read_acs5year_estimate_margin_}
\alias{read_acs5year_estimate_margin_}
\title{Read a ACS 5-year survey estimate of a state (internal)}
\usage{
read_acs5year_estimate_margin_(path_to_census, state, year, file_seg,
  est_marg = "e", show_progress = TRUE)
}
\arguments{
\item{path_to_census}{path to the directory of downloaded census 2010 data,
inside which are the subfolders of each state.}

\item{state}{abbrivation of a state, for example, "IN" for Indiana.}

\item{year}{end year of the 5-year survey}

\item{file_seg}{the number of the data file, for example, "0012.}

\item{est_marg}{which data to read, "e" for estimate, "m" for margin}
}
\value{
A data.table of all data in the data file. LOGRECNO is its key.
}
\description{
Read an entire ACS 5-year survey estimate file in a state and assign correct reference
to each column. Search with \code{\link{search_acs_datafile}} to find table contents
in a data file.
}
\examples{
\dontrun{
# read Rhode Island data file 02
ri <- read_acs5year_estimate_margin_("~/census_data/", "ri", 2015, "group2",
                                     "0019", "e")
}

}
\seealso{
\code{\link{dict_acs_datafile}}
}
