% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_censustable.R
\name{read_censustable}
\alias{read_censustable}
\title{Read selected census table}
\usage{
read_censustable(path_to_census, state, table_num)
}
\arguments{
\item{path_to_census}{path to the directory holding downloaded census data}

\item{state}{abbrivation of a state, such as "IN" for Indiana}

\item{table_num}{table number such as "P3" and "PCT11A"}
}
\description{
Read a census table in a state
}
\examples{


}
