% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_datafile.R
\name{read_datafile}
\alias{read_datafile}
\title{Read data file}
\usage{
read_datafile(path_to_census, state, file_num)
}
\arguments{
\item{path_to_census}{path to the directory of downloaded census 2010 data}

\item{state}{abbrivation of a state, for example, "IN" for Indiana}

\item{file_num}{the number of the file, for example, 5 for file 05}
}
\description{
Read an entire data file in a state and assign correct reference to each column
}
\examples{


}
\seealso{
\code{\link{dict_datafile}}
}
