% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_census.R
\name{read_decennial}
\alias{read_decennial}
\title{Read selected geographic headers and table contents of selected summary level
and geographic component.}
\usage{
read_decennial(year, states, table_contents = NULL, geo_headers = NULL,
  summary_level = "*", geo_comp = "*", with_coord = TRUE,
  with_population = TRUE, show_progress = TRUE)
}
\arguments{
\item{year}{year of the census}

\item{table_contents}{selected references of contents in census tables.}

\item{summary_level}{select which summary level to keep, "*" to keep all. It takes strings
including "state", "county", "county_subdivision", "place", "tract", "block_group",
and "block" for the most common levels. It also take code for level. Search all codes with
\code{\link{search_sumlev}} for state files or \code{\link{search_sumlev_US}} for national files.}

\item{geo_comp}{select which geographic component to keep, "*" to keep all.
Summary level of a state contains all geographic component. County
subdivision and higher level have all area ("00"), urban ("01"), and rural ("43").
census tract and lower level have no geographic component, all are "00". Browse
\code{\link{dict_geocomp}} for other geographic components.}

\item{show_progress}{show progress of file reading if TRUE. Turn off if FALSE, which
is useful in RMarkdown output.}

\item{state}{vector of abbreviation of states, for example "IN" or c("MA", "RI").}

\item{geoheaders}{vector of references of selected geographci headers to be included in the return.}
}
\value{
A data.table whose columns include the selected geoheaders and
table contents plus LOGRECNO, SUMLEV, GEOCOMP, and state. state indicates
the data from state files with state abbreviation or national files with "US".
}
\description{
Read geographic header record file and data files of states and return
selected geoheaders, state abbreviation, and table contents of selected summary
level and geographic component. These table contents can be from different tables, for example,
c("PCT012F139", "P0030008", "P0100059", "P0150008", "H0070016", "PCT012F138").
The rows of short contents that have no corresponding LOGRECNOs are filled with
NAs. To find the references of table contents of interest, search with function
\code{\link{search_datafile}}. To find geographic headers, browse
\code{\link{dict_geoheader}} or search with function \code{\link{search_geoheader}}.
}
\examples{
\dontrun{
# read Rhode Island and Massachusette summary data at block level
aaa = read_decennial(year = 2010,
                     states = c("ut", "ri"),
                     table_contents = c("P0030002", "P0150008"),
                     geo_headers = c("place = ut62360",  # place in UT
                                     "place = 14140",    # place in UT or RI
                                     "cousub = ri41500"),
                     summary_level = "block")
}

}
