% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_field.R
\name{read_fields}
\alias{read_fields}
\title{Read data from selected fields}
\usage{
read_fields(path_to_census, state, fields)
}
\arguments{
\item{path_to_census}{path to the directory holding downloaded census data}

\item{state}{abbrivation of a state, such as "IN" for Indiana}

\item{fields}{selected field references in census tables}
}
\description{
These fields can be from different tables but should have same number of rows,
for example, c("PCT012F139", "P0030008", "P0100059", "P0150008", "H0070016", "PCT012F138")
}
\examples{
\dontrun{
path <- your_local_path_to_census_data
fields <- c("PCT012F139", "P0030008", "P0100059", "P0150008", "H0070016", "PCT012F138")
dt <- read_fields(path, "RI", fields)

}

}
