% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geographics.R
\name{read_geographics}
\alias{read_geographics}
\title{Read geographic header record file of a state and return logical record number
and selected fields}
\usage{
read_geographics(path_to_census, state, field, show_progress = TRUE)
}
\arguments{
\item{path_to_census}{path to the directory holding downloaded
census 2010 summary file 1 with urban/rural update}

\item{state}{abbreviation of a state, for example "IN" for "Indiana"}

\item{field}{vector of references of selected geographci headers to be included in the return}
}
\value{
data.table whose columns are logical record number and selected fields
}
\description{
Read geographic header record file of a state and return logical record number
and selected fields
}
\examples{

}
