% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_table.R
\name{read_table}
\alias{read_table}
\title{Read selected table}
\usage{
read_table(path_to_census, state, table_num)
}
\arguments{
\item{path_to_census}{path to the directory holding downloaded census data}

\item{state}{abbrivation of a state, such as "IN" for Indiana}

\item{table_num}{table number such as "P3" and "PCT11A"}
}
\description{
Read a table in a state
}
\examples{


}
