% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tablecontent.R
\name{read_tablecontents}
\alias{read_tablecontents}
\title{Read data of selected table fields}
\usage{
read_tablecontents(path_to_census, state, contents)
}
\arguments{
\item{path_to_census}{path to the directory holding downloaded census data}

\item{state}{abbrivation of a state, such as "IN" for Indiana}

\item{contents}{selected references of contents in census tables}
}
\value{
A data.table containing "LOGRECNO" and selected table contents
}
\description{
These fields can be from different tables, for example,
c("PCT012F139", "P0030008", "P0100059", "P0150008", "H0070016", "PCT012F138").
The rows of short fields that have no corresponding LOGRECNOs are filled with
NAs.
}
\examples{
\dontrun{
path <- your_local_path_to_census_data
contents <- c("PCT012F139", "P0030008", "P0100059", "P0150008", "H0070016", "PCT012F138")
dt <- read_tablecontents(path, "RI", contents)

}

}
