% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_county_fips_dictionary.R
\name{search_countyfips}
\alias{search_countyfips}
\title{Search county fips}
\usage{
search_countyfips(keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword to search in fips or county name. To search for fips
contains both words "abc" and "defg", the keyword is simply a single string
of "abc defg".}

\item{view}{display the search result with View if TRUE}
}
\description{
search census tables by keyword in table numbers or table descriptions
}
\examples{

}
\seealso{
\code{\link{dict_countyfips}} lists all county fips
}
