% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fips.R
\name{search_fips}
\alias{search_fips}
\title{Search FIPS}
\usage{
search_fips(keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword to be searched}

\item{view}{display the search result with View if TRUE}
}
\value{
A data.table
}
\description{
Search FIPS
}
\examples{
\dontrun{
  # search fips of Lincoln in Rhode Island
  search_fips("rhode island lincoln")

  # list fips of all counties in Massachusetts, even cannot spell correctly
  search_fips("massachu county")
}
}
\seealso{
\code{\link{dict_fips}}
}
