% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_geographic_component.R
\name{search_geocomp}
\alias{search_geocomp}
\title{Search geographic components}
\usage{
search_geocomp(keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword to search in code or description}

\item{view}{display the search result with View if TRUE}
}
\value{
A data.table
}
\description{
Search the code or content of geographic components in dataset
\code{\link{dict_geocomp}}.
}
\details{
The most frequently used geographic components are:

00 : all geographic component

01 : urban

43 : rural
}
\examples{
\dontrun{
  # search geocomponents containing "urban cluster"
  search_geocomp("urban cluster")

  # search geocomponents with code 43
  search_geocomp("43")
  # or
  search_geocomp(43)
}

}
\seealso{
\code{\link{dict_geocomp}} lists all geocomponents and codes
}
