% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geographics.R
\name{search_geodict}
\alias{search_geodict}
\title{Search for field reference with a keyword or search for the description of a
a field from its reference.}
\usage{
search_geodict(ref = NULL, keyword = NULL)
}
\arguments{
\item{ref}{reference of a field such as "SUMLEV"}

\item{keyword}{keyword in description}
}
\value{
data.table match the search criteria
}
\description{
Search for field reference with a keyword or search for the description of a
a field from its reference.
}
\examples{

}
