% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_geographic_header.R
\name{search_geoheader}
\alias{search_geoheader}
\title{Search geographic headers}
\usage{
search_geoheader(keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword in description or reference}

\item{view}{display the search result with View if TRUE}
}
\value{
data.table matching the search criteria
}
\description{
Search in field reference or description of geographic header
with keyword in dataset \code{\link{dict_geoheader}}.
}
\examples{
\dontrun{
  # search geoheader that contains keyword "india"
  search_geoheader("india")

  # search for lattitude
  search_geoheader("latitu")

  # search geoheader of reference containing "INTPT"
  search_geoheader("INTPT")
}


}
