% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_geographic_header_dictionary.R
\name{search_geoheader_dict}
\alias{search_geoheader_dict}
\title{Search in field reference or description with a keyword}
\usage{
search_geoheader_dict(keyword)
}
\arguments{
\item{keyword}{keyword in description or reference}
}
\value{
data.table match the search criteria
}
\description{
Search in field reference or description with a keyword
}
\examples{

}
