% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_census_and_acs.R
\name{search_geoheaders}
\alias{search_geoheaders}
\title{Search geographic headers}
\usage{
search_geoheaders(survey, keyword = "*", view = TRUE)
}
\arguments{
\item{survey}{type of survey, either "census" or "acs".}

\item{keyword}{keyword in description or reference. The default "*" includes
all geoheaders.}

\item{view}{display the search result with View if TRUE}
}
\value{
data.table matching the search criteria
}
\description{
Search in field reference or description of geographic header
with keyword in dataset \code{\link{dict_census_geoheader}} or
\code{\link{dict_acs_geoheader}}.
}
\examples{
\dontrun{
  # search geoheader that contains keyword "india" in Census 2010
  search_geoheaders("census", "india")

  # search for lattitude
  search_geoheaders("census", "latitu")

  # browse all geoheaders in ACS
  search_geoheaders("acs")
}

}
