% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_summary_level_dictionary.R
\name{search_sumlev}
\alias{search_sumlev}
\title{Search summary level}
\usage{
search_sumlev(keyword)
}
\arguments{
\item{keyword}{keyword to search in code or description}
}
\description{
search the code or description of summary level
}
\examples{
# search geocomponent contains "block"
search_sumlev("block")

# search geocomponent with code 40
search_sumlev("40")
# or
search_sumlev(40)
}
\seealso{
\code{\link{summarylevel_dict}} lists all geocomponents and codes
}
