% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_summary_level_national_dictionary.R
\name{search_sumlev_US}
\alias{search_sumlev_US}
\title{Search summary level}
\usage{
search_sumlev_US(keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword to search in code or description}

\item{view}{display the search result with View if TRUE}
}
\description{
search the code or description of summary level
}
\examples{
# search geocomponent contains "block"
search_sumlev("block")

# search geocomponent with code 40
search_sumlev("40")
# or
search_sumlev(40)
}
\seealso{
\code{\link{dict_summarylevel_US}} lists all geocomponents and codes
}
