% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_census_and_acs.R
\name{search_summarylevels}
\alias{search_summarylevels}
\title{Search summary levels}
\usage{
search_summarylevels(survey, keyword = "*", view = TRUE)
}
\arguments{
\item{survey}{"decennial" or "acs"}

\item{keyword}{keyword to search in code or description}

\item{view}{display the search result with View if TRUE}
}
\value{
A data.table of searched results.
}
\description{
Search code or description of summary levels
}
\examples{
# Change view = TRUE (default) to View the returned data.
aaa = search_summarylevels("decennial", "block", view = FALSE)
bbb <- search_summarylevels("acs", "40", view = FALSE)

\dontrun{
  # view all summary levels
  search_summarylevels("decennial")
  search_summarylevels("acs")
}

}
