% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_census_table_dictionary.R
\name{search_table}
\alias{search_table}
\title{Search census table}
\usage{
search_table(keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword to search in code or description. To search for a table
contains two words "abc" and "defg", the keyword is simply a single string
of "abc defg".}

\item{view}{display the search result with View if TRUE}
}
\description{
search census tables by keyword in table numbers or table descriptions
}
\examples{
# search census table contains "occupancy"
search_table("occupancy")

# search census table with table number "H6"
search_table("H6")
}
\seealso{
\code{\link{dict_censustable}} lists all geocomponents and codes
}
