% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_census_and_acs.R
\name{search_tablecontents}
\alias{search_tablecontents}
\title{Search table contents in data files}
\usage{
search_tablecontents(survey, keyword, year = NULL, view = TRUE)
}
\arguments{
\item{survey}{either "decennial" for decenial or "acs" or American Community Survey.}

\item{keyword}{keyword to be searched}

\item{year}{ending year of the survey}
}
\value{
A data.table
}
\description{
Search in lookup datasets of each survey to find references of
for table_contents argument in function \code{\link{read_decennial2010}},
\code{\link{read_acs1year}}, and \code{\link{read_acs5year}}.
}
\examples{
\dontrun{
  # search exactly table "p3"
  search_datafile("p3", table_only = TRUE)

  # search reference "p0030002"
  search_datafile("p0030002")

  # search for file 03
  search_datafile("03", file_only = TRUE)

  # search keyword "federal prision population".
  search_datafile("federal prison population")
}

}
