% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_census_and_acs.R
\name{search_tables}
\alias{search_tables}
\title{search decennial and acs tables by keyword in table numbers or table descriptions}
\usage{
search_tables(survey, keyword = "*", view = TRUE)
}
\arguments{
\item{survey}{"decennial" or "acs"}

\item{keyword}{keyword to search in code or description.}

\item{view}{display the search result with View if TRUE}
}
\value{
A data.table
}
\description{
search decennial and acs tables by keyword in table numbers or table descriptions
}
\examples{
# Change view = TRUE (default) to View the returned data.
aaa <- search_tables("decennial", "occupancy", view = FALSE)
bbb <- search_tables("acs", "detailed race", view = FALSE)

\dontrun{
  # view all tables
  search_tables("decennial")
  search_tables("acs")
}


}
