% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_states_DC.R
\docType{data}
\name{states_DC}
\alias{states_DC}
\title{Vector of the abbreviations of 50 states and DC}
\format{A vector of 51 element}
\usage{
data("states_DC")
}
\description{
Abbrivation only
}
\keyword{datasets}
