% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summarylevel_dict.R
\docType{data}
\name{summarylevel_dict}
\alias{summarylevel_dict}
\title{Summary levels and codes}
\format{data.table}
\usage{
data("geo_component")
}
\description{
This data serves as a dictionary for summary level codes and their
discription for summary file 1 with urban/rural update.
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 4-16 state summary file with urban/rural update
}
\keyword{datasets}
