% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{table_content_acs1year_all_years}
\alias{table_content_acs1year_all_years}
\title{ACS 1-year table contents of all years}
\format{
A data.table with 27246 rows and 7 variables
\describe{
  \item{reference}{reference of the table content, such as "B01001_002". The reference
        is used to extract data of table content.}
  \item{table_content}{description of columns in a table}
  \item{table_name}{table names}
  \item{acs1_2019}{restriction and availability of table content in 2019}
  \item{acs1_2018}{restriction and availability of table content in 2018}
  \item{acs1_2017}{restriction and availability of table content in 2017}
  \item{acs1_2017}{restriction and availability of table content in 2016}
  \item{acs1_2015}{restriction and availability of table content in 2015}
  \item{acs1_2014}{restriction and availability of table content in 2014}
  \item{acs1_2013}{restriction and availability of table content in 2013}
  \item{acs1_2012}{restriction and availability of table content in 2012}
  \item{acs1_2011}{restriction and availability of table content in 2011}
  \item{acs1_2010}{restriction and availability of table content in 2010}
  \item{acs1_2009}{restriction and availability of table content in 2009}
  \item{acs1_2008}{restriction and availability of table content in 2008}
  \item{acs1_2007}{restriction and availability of table content in 2007}
  \item{acs1_2006}{restriction and availability of table content in 2006}
  \item{acs1_2005}{restriction and availability of table content in 2005}
  \item{uiverse}{the universe of the data}
}
}
\source{
Check for each year of ACS 1-year and 5-year
\href{https://www.census.gov/programs-surveys/acs/technical-documentation/summary-file-documentation.2015.html}{Sequence Number/Table Number Lookup File}.
}
\usage{
table_content_acs1year_all_years
}
\description{
There is slightly difference in the table contents of each year.
}
\keyword{datasets}
