% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{dict_acs1_geocomponent}
\alias{dict_acs1_geocomponent}
\title{List of summary levels used in ACS 5 year surverys}
\format{A data.table with 19 rows and 6 variables:
\describe{
  \item{code}{code for the geocomponent, such as "01" and "M3"}
  \item{geo_component}{description of the geographic component}
  \item{state_acs1}{wheather a geocomponent available in 1-year survey state files}
  \item{US_acs1}{wheather a geocomponent available in 1-year survey national files}
  \item{state_acs5}{wheather a geocomponent available in 5-year survey state files}
  \item{US_acs5}{wheather a geocomponent available in 5-year survey national files}
}}
\usage{
dict_acs1_geocomponent
}
\description{
List of summary levels used in ACS 5 year surverys
}
\keyword{datasets}
