% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R, R/data_decennial.R
\docType{data}
\name{dict_all_geocomponent_2010}
\alias{dict_all_geocomponent_2010}
\alias{dict_all_geocomponent_2010}
\title{List of all geographic components, 2010 version}
\format{A data.table with 114 rows and 2 variables:
\describe{
  \item{code}{code for the geocomponent, such as "01" and "M3"}
  \item{geo_component}{description of the geographic component}
}}
\source{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 6-15
}
\usage{
dict_all_geocomponent_2010

dict_all_geocomponent_2010
}
\description{
List of all geographic components, 2010 version

This dataset contains all available geographic components and codes.
}
\keyword{datasets}
