% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_censustable_dict.R
\docType{data}
\name{dict_censustable}
\alias{dict_censustable}
\title{Complete list of census tables}
\format{A data.table with 333 rows and 4 variables:
\describe{
  \item{table_ref}{reference code such as "H0010", "PCT022G"}
  \item{table_number}{table number such as "H1", "PCT22G"}
  \item{table_name}{description of the table}
  \item{universe}{unit of the table}
}}
\usage{
data("dict_censustable")
}
\description{
All census tables in census 2010 summary file 1 with urban/rural update
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
all across chapter 6.
}
\seealso{
\code{\link{search_table}}
}
\keyword{datasets}
