% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geocomponent_dict.R
\docType{data}
\name{dict_geocomp}
\alias{dict_geocomp}
\title{List of geographic components and codes}
\format{A data.table with 114 rows and 2 variables:
\describe{
  \item{code}{code for the geocomponent, such as "01" and "M3"}
  \item{geo_component}{description of the geocomponent}
}}
\usage{
data("dict_geocomp")
}
\description{
List of geographic components and codes
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 6-15
}
\seealso{
\code{\link{search_geocomp}}
}
\keyword{datasets}
