% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geoheader_dict.R
\docType{data}
\name{dict_geoheader}
\alias{dict_geoheader}
\title{Complete list of geographic Header Record in Summary File 1}
\format{A data.table with 101 rows and 4 variables
\describe{
  \item{reference}{reference of the geoheader record}
  \item{description}{description of the geoheader}
  \item{start}{starting position of the geoheader in the record}
  \item{end}{ending position of the geoheader in the record}
}}
\usage{
data("dict_geoheader")
}
\description{
This data serves as a dictionary for geographic header references and their
discription.
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 2-8
}
\seealso{
\code{\link{search_geoheader}}
}
\keyword{datasets}
