% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summarylevel_dict.R
\docType{data}
\name{dict_summarylevel}
\alias{dict_summarylevel}
\title{Summary levels and codes for state files}
\format{A data.table with 99 rows and 2 variables
\describe{
  \item{code}{code of summary level}
  \item{description}{description of summary level}
}}
\usage{
data("dict_summarylevel")
}
\description{
This data serves as a dictionary for summary level codes and their
discription for summary file 1 with urban/rural update.
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 4-16 state summary file with urban/rural update
}
\seealso{
\code{\link{search_sumlev}}
}
\keyword{datasets}
