% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summarylevel_national_dict.R
\docType{data}
\name{dict_summarylevel_US}
\alias{dict_summarylevel_US}
\title{Summary levels and codes for national files}
\format{A data.table with 77 rows and 2 variables
\describe{
  \item{code}{code of summary level}
  \item{description}{description of summary level}
}}
\usage{
data("dict_summarylevel_US")
}
\description{
This data serves as a dictionary for summary level codes and their
discription for summary file 1 with urban/rural update.
}
\references{
2010 Census Summary File 1
\href{https://www.census.gov/prod/cen2010/doc/sf1.pdf}{technical documentation}
page 4-25 state summary file with urban/rural update
}
\seealso{
\code{\link{search_sumlev_US}}
}
\keyword{datasets}
