% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_census.R
\name{download_census}
\alias{download_census}
\title{download census data}
\usage{
download_census(survey = NULL, year = NULL, states = c(states_DC, "US",
  "PR"))
}
\arguments{
\item{survey}{Which survey to download from, "decennial", "acs5year", or "acs1year"}

\item{year}{year or ending year of the survey}

\item{states}{vector of abbreviations of states such as c("MA", "RI")}
}
\description{
Download census data from United States Census bureau. By default
it downloads summary files and extract summary files of Census 2010, ACS 5-year
survey of 2015, and ACS 1-year survey of 2016, 2015, and 2014. It also download
generated data from Census 2010 if not exist.
}
