% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_geoid_coordinate.R
\name{generate_geoidcoord}
\alias{generate_geoidcoord}
\title{Generate coordinate for all GEOIDs in ACS from Census 2010.}
\usage{
generate_geoidcoord(states = c(states_DC, "PR", "US"))
}
\description{
It creates a sub directory "generared_data/" under path_to_census
if not exsist and a file called geoid_coord.csv. The coordinates are generated
from census 2010 summary file 1 (with urban/rural update). GEOIDs include all
those in 2015 ACS 1-year and 5-year survey. The coordinate of
new GEOID created after Census 2010 are not available.
}
