% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acs.R
\docType{data}
\name{lookup_acs1year_2016}
\alias{lookup_acs1year_2016}
\title{ACS 1-year 2016 file segment and table lookup data}
\format{A data.table with 31835 rows and 6 variables
\describe{
  \item{file_segment}{sequence number of segment data files, from "0001" to "0166"}
  \item{table_content}{description of columns in a table}
  \item{reference}{reference of the table content, such as "B01001_002". The reference
        is used to extract data of table content.}
  \item{table_number}{table number such as "B01001"}
  \item{table_name}{description of table. A table has multiple columns (table_content)}
  \item{universe}{the universe of the data}
}}
\source{
Check for each year of ACS 1-year and 5-year
\href{https://www.census.gov/programs-surveys/acs/technical-documentation/summary-file-documentation.2015.html}{Sequence Number/Table Number Lookup File}.
}
\usage{
lookup_acs1year_2016
}
\description{
There is slightly difference in the lookup tables of each year.
}
\keyword{datasets}
