% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_census.R
\name{read_2010census}
\alias{read_2010census}
\title{Read selected geographic headers and table contents}
\usage{
read_2010census(path_to_census, state, geoheaders, table_contents,
  show_progress = TRUE)
}
\arguments{
\item{path_to_census}{path to the directory holding downloaded
census 2010 summary file 1 (with urban/rural update). Inside this directiory
are sub-folders of each state.}

\item{state}{abbreviation of a state, for example "IN" for "Indiana".}

\item{geoheaders}{vector of references of selected geographci headers to be included in the return.}

\item{table_contents}{selected references of contents in census tables.}

\item{show_progress}{show progress of file reading if TRUE. Turn off if FALSE, which
is useful in RMarkdown output.}
}
\value{
A data.table whose columns are LOGRECNO, SUMLEV, GEOCOMP, the selected geoheaders and
the table contents, with LOGRECNO as the key. The data.table contains all rows
of geoheader and data files.
}
\description{
Read geographic header record file and data files of a state and return logical record number,
selected geoheaders and table contents. These table contents can be from different tables, for example,
c("PCT012F139", "P0030008", "P0100059", "P0150008", "H0070016", "PCT012F138").
The rows of short contents that have no corresponding LOGRECNOs are filled with
NAs. To find the references of table contents of interest, search with function
\code{\link{search_datafile}}. To find geographic headers, browse
\code{\link{dict_geoheader}} or search with function \code{\link{search_geoheader}}.
}
\examples{
\dontrun{
# read Rhode Island summary data
ri <- read_2010census("your_local_path_to_2010_census", "RI",
                       c("GEOCOMP", "NAME", "SUMLEV"),
                       c("PCT012F139", "PCT012F138", "P0150008", "H0070016"))

# read the national summary data
us <- read_2010census("your_local_path_to_2010_census", "US",
                       c("STATE", "COUNTY", "COUSUB", "GEOCOMP", "NAME", "SUMLEV"),
                       c("PCT012F139", "PCT012F138", "P0150008", "H0070016"))
}

}
