% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acs1year.R
\name{read_acs1year}
\alias{read_acs1year}
\title{Read ACS 1-year survey}
\usage{
read_acs1year(year, states, table_contents = NULL, geo_headers = NULL,
  summary_level = "*", geo_comp = "*", with_margin = FALSE,
  with_coord = TRUE, with_population = TRUE, with_acsgeoheaders = FALSE,
  show_progress = TRUE)
}
\arguments{
\item{year}{year of the 1-year survey}

\item{states}{vector of abbreviations of states such as c("MA", "RI")}

\item{table_contents}{vector of reference of available table contents}

\item{geo_headers}{vector of geographic headers like c("NAME", "PLACE")}

\item{summary_level}{summary level like "050"}

\item{geo_comp}{geographic component such as "00", "01", and "43"}

\item{with_margin}{read also margin of error in addition to estimate}

\item{with_coord}{whether to include longitude and latitude in return}

\item{with_population}{whether to include total population in return}

\item{with_acsgeoheaders}{whether to include geographic headers from ACS data}

\item{show_progress}{whether to show progress in fread()n}
}
\value{
A data.table of selected data.
}
\description{
Read ACS 1-year survey of selected geography, estimate, and margin
of selected states.
}
\examples{
\dontrun{
# The lowest summary level in ACS 1-year is county subdivision. There is no
# need to run fine search in geo_headers but just use NULL geo_headers.
# The example get data of all counties in UT and RI.
aaa <- read_acs1year(2015, c("ut", "ri"),
                      table_contents = c("B01001_009", "B00001_001", "B10001_002"),
                      summary_level = "county")
}


}
