% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acs_geography_file.R
\name{read_acs1year_geo_}
\alias{read_acs1year_geo_}
\title{Read ACS 1-year survey geography file of a state}
\usage{
read_acs1year_geo_(path_to_census, state, year, references = c("NAME",
  "GEOID"), show_progress = TRUE)
}
\arguments{
\item{path_to_census}{path to the directory holding downloaded ACS files.}

\item{state}{abbreviation of a state, for example "IN" for "Indiana".}

\item{year}{year of the survey}

\item{references}{vector of references of selected geographic headers to be included in the return}

\item{show_progress}{show progress of reading if TRUE. Turn off if FALSE, which
is useful in RMarkdown output.}
}
\value{
data.table whose columns are logical record number, summary level,
geographic compoenent, and selected references. LOGRECNO serves as
the key.
}
\description{
Read ACS 1-year survey geography file of a state and return
logical record number, summary level, geographic component
and selected geographic headers. To find geographic headers, browse
\code{\link{dict_acs_geoheaders}} or search with \code{\link{search_acs_geoheader}}.
}
\examples{
\dontrun{
# read selected geographic headers of Rhode Island
ri  <- read_acs1year_geo_("your_local_path_to_census_data", "RI", 2015,
                      c("NAME", "STATE"))
}

}
