% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_acs.R
\name{read_acs5year}
\alias{read_acs5year}
\title{Read ACS 5-year survey}
\usage{
read_acs5year(path_to_census, states, year, geo_headers = NULL,
  table_contents = NULL, with_margin = TRUE, with_coord = TRUE,
  summary_level = "*", geo_comp = "*", show_progress = TRUE)
}
\arguments{
\item{path_to_census}{path to the directory of downloaded census data.}

\item{states}{vector of abbrivations of states, for example, "IN" and c("MA", "RI").}

\item{year}{year of the survey}

\item{geo_headers}{geographic headers.}

\item{est_marg}{which data to read, "e" for estimate, "m" for margin}
}
\value{
A data.table of selected data.
}
\description{
Read ACS 5-year survey of selected geography, estimate, and margin
of selected states.
}
\examples{
\dontrun{
# read Rhode Island data file 0004, replace the path with your local folder
ma_ri <- read_acs1year("~/census_data", c("ma", "ri"), 2015,
                    geo_headers = c("NAME", "GEOID"),
                    table_contents = c("B01001_009", "B00001_001", "B10001_002"),
                    summary_level = "place")
}


}
