% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geographic_header_record_file.R
\name{read_geoheader}
\alias{read_geoheader}
\title{Read geographic header record file}
\usage{
read_geoheader(path_to_census, state, field, show_progress = TRUE)
}
\arguments{
\item{path_to_census}{path to the directory holding downloaded
census 2010 summary file 1 with urban/rural update}

\item{state}{abbreviation of a state, for example "IN" for "Indiana"}

\item{field}{vector of references of selected geographci headers to be included in the return}

\item{show_progress}{show progress of reading if TRUE. Turn off if FALSE, which
is useful in RMarkdown output.}
}
\value{
data.table whose columns are logical record number and selected fields
}
\description{
Read geographic header record file of a state and return logical record number
and selected fields
}
\examples{
\dontrun{
path <- locat_path_to_census_data
selected_geoheader <- read_geoheader(
    path_to_census = path,
    state = "RI",
    field = c("SUMLEV", "INTPTLAT", "INTPTLON")
)
}


}
