% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_datafile.R
\name{search_datafile}
\alias{search_datafile}
\title{Search data files}
\usage{
search_datafile(keyword, table_only = FALSE, file_only = FALSE,
  view = TRUE)
}
\arguments{
\item{keyword}{keyword to be searched}

\item{table_only}{only search in and strictly match table_num}

\item{file_only}{only search in file numbers}

\item{view}{display the search result with View if TRUE}
}
\value{
A data.table
}
\description{
Search in all data files in order to find the file containing
the keyword.
}
\examples{
\dontrun{
  # search exactly table "p3"
  search_datafile("p3", table_only = TRUE)

  # search reference "p0030002"
  search_datafile("p0030002")

  # search file 03
  search_datafile("03", file_only = TRUE)

  # search keyword "asian total population". too many matches, usually do not do it
  search_datafile("asian total population")
}

}
\seealso{
\code{\link{dict_datafile}}
}
