% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_datafile.R
\name{search_datafile}
\alias{search_datafile}
\title{Search in data files}
\usage{
search_datafile(keyword, table_only = FALSE, file_only = FALSE,
  view = TRUE)
}
\arguments{
\item{keyword}{keyword to be searched}

\item{table_only}{only search in and strictly match table number}

\item{file_only}{only search in file numbers}

\item{view}{display the search result with View if TRUE}
}
\value{
A data.table
}
\description{
Search in dataset \code{\link{dict_datafile}} to find the file
or table containing the keywords. This function helps users to find what data are
included in the summary file 1 (with urban/rural update).
}
\examples{
\dontrun{
  # search exactly table "p3"
  search_datafile("p3", table_only = TRUE)

  # search reference "p0030002"
  search_datafile("p0030002")

  # search for file 03
  search_datafile("03", file_only = TRUE)

  # search keyword "federal prision population".
  search_datafile("federal prison population")
}

}
