% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fips.R
\name{search_fips}
\alias{search_fips}
\title{Search fips}
\usage{
search_fips(keyword, table_only = FALSE, view = TRUE)
}
\arguments{
\item{keyword}{keyword to be searched}

\item{view}{display the search result with View if TRUE}
}
\description{
Search fips
}
\examples{


}
\seealso{
\code{\link{dict_fips}}
}
