% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_geographic_component_dictionary.R
\name{search_geocomp}
\alias{search_geocomp}
\title{Search geographic component}
\usage{
search_geocomp(keyword)
}
\arguments{
\item{keyword}{keyword to search in code or description}
}
\description{
search the code or description of geographic component
}
\examples{
# search geocomponent contains "urban cluster"
search_geocomp("urban cluster")

# search geocomponent with code 43
search_geocomp("43")
# or
search_geocomp(43)
}
\seealso{
\code{\link{geocomp_dict}} lists all geocomponents and codes
}
