% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_geographic_component_dictionary.R
\name{search_geocomp}
\alias{search_geocomp}
\title{Search geographic component}
\usage{
search_geocomp(keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword to search in code or description}

\item{view}{display the search result with View if TRUE}
}
\description{
search the code or description of geographic component
}
\examples{
# search geocomponent contains "urban cluster"
search_geocomp("urban cluster")

# search geocomponent with code 43
search_geocomp("43")
# or
search_geocomp(43)
}
\seealso{
\code{\link{dict_geocomp}} lists all geocomponents and codes
}
