% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_geographic_header_dictionary.R
\name{search_geoheader}
\alias{search_geoheader}
\title{Search geographic header}
\usage{
search_geoheader(keyword)
}
\arguments{
\item{keyword}{keyword in description or reference}
}
\value{
data.table match the search criteria
}
\description{
Search in field reference or description with a keyword
}
\examples{
# search geoheader that contains keyword "india"
search_geoheader("india")

# search geoheader of reference containing "INTPT"
search_geoheader("INTPT")

}
\seealso{
\code{\link{geoheader_dict}}, which lists all the geographic headers
}
