% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_geographic_header_dictionary.R
\name{search_geoheader}
\alias{search_geoheader}
\title{Search geographic headers}
\usage{
search_geoheader(keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword in description or reference}

\item{view}{display the search result with View if TRUE}
}
\value{
data.table matching the search criteria
}
\description{
Search in field reference or description with a keyword
}
\examples{
\dontrun{
  # search geoheader that contains keyword "india"
  search_geoheader("india")

  # search geoheader of reference containing "INTPT"
  search_geoheader("INTPT")
}

}
\seealso{
\code{\link{dict_geoheader}}, which lists all the geographic headers
}
