% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_summary_level.R
\name{search_sumlev}
\alias{search_sumlev}
\title{Search summary levels of state files}
\usage{
search_sumlev(keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword to search in code or description}

\item{view}{display the search result with View if TRUE}
}
\value{
A data.table of searched results.
}
\description{
Search code or description of summary levels used in state files.
The summary levels are stored in dataset \code{\link{dict_summarylevel}}.
This dataset is different from that for national files,
\code{\link{dict_summarylevel_US}}. Function \code{\link{search_sumlev_US}}
searches summary levels in national files.
}
\examples{
\dontrun{
  # search summary levels of geocomponent contains "block"
  search_sumlev("block")

  # search summary levels of code 40
  search_sumlev("40")
}

}
