% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_summary_level_dictionary.R
\name{search_sumlev}
\alias{search_sumlev}
\title{Search summary levels of state files}
\usage{
search_sumlev(keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword to search in code or description}

\item{view}{display the search result with View if TRUE}
}
\value{
A data.table of searched results.
}
\description{
Search the code or description of summary levels of state files.
}
\examples{
# search summary levels of geocomponent contains "block"
search_sumlev("block")

# search summary levels of code 40
search_sumlev("40")
# or
search_sumlev(40)

}
\seealso{
\code{\link{dict_summarylevel}} lists all geocomponents and codes
}
