% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_census_and_acs.R
\name{search_summarylevels}
\alias{search_summarylevels}
\title{Search summary levels}
\usage{
search_summarylevels(survey, keyword = "*", view = TRUE)
}
\arguments{
\item{survey}{"decennial" or "acs"}

\item{keyword}{keyword to search in code or description}

\item{view}{display the search result with View if TRUE}
}
\value{
A data.table of searched results.
}
\description{
Search code or description of summary levels
}
\examples{
\dontrun{
  # search summary levels of geocomponent contains "block"
  search_sumlev("block")

  # search summary levels of code 40
  search_sumlev("40")
}

}
