% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_census_and_acs.R
\name{search_table}
\alias{search_table}
\title{Search decennial tables}
\usage{
search_table(survey, keyword, view = TRUE)
}
\arguments{
\item{keyword}{keyword to search in code or description. To search for a table
contains two words "abc" and "defg", the keyword is simply a single string
of "abc defg".}

\item{view}{display the search result with View if TRUE}
}
\value{
A data.table
}
\description{
search decennial tables by keyword in table numbers or table descriptions
}
\examples{
\dontrun{
  # search decennial table contains "occupancy"
  search_table("occupancy")

  # search decennial table with table number "H5"
  search_table("H5")
}

}
\seealso{
\code{\link{dict_decennialtable}} lists all geocomponents and codes
}
