% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_path_to_census.R
\name{set_path_to_census}
\alias{set_path_to_census}
\title{Set file path to directory holding census data}
\usage{
set_path_to_census(path)
}
\arguments{
\item{path}{path to directory holding all downloaded census data, such as
"E:/my_census_data" and "~/my_census_data/".}
}
\description{
Set file path to directory holding census data
}
