% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_path_to_census.R
\name{set_path_to_census}
\alias{set_path_to_census}
\title{Set file path to directory holding census data}
\usage{
set_path_to_census(path, overwrite = TRUE, install = TRUE)
}
\arguments{
\item{path}{path to directory holding all downloaded census data, such as
"E:/my_census_data/" and "~/my_census_data/".}

\item{overwrite}{whether to overwrite old .Renviron.}

\item{install}{whether to set the path for all future use.}
}
\description{
Set file path to directory holding census data
}
